Die verschiedenen Screens ausgelagert! // Wichtig fuer den Navigator zum wechseln der einzelnen Screen
